% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_files}
\alias{add_files}
\title{Add files to GitAI object.}
\usage{
add_files(gitai, files)
}
\arguments{
\item{gitai}{A GitAI object.}

\item{files}{A character vector of file paths. May be defined with
regular expression.}
}
\value{
GitAI object.
}
\description{
Add files to GitAI object.
}
