% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_records.R
\name{find_records}
\alias{find_records}
\title{Finding top K records in a vector database.}
\usage{
find_records(gitai, query, top_k = 1, verbose = is_verbose())
}
\arguments{
\item{gitai}{A \code{GitAI} object.}

\item{query}{A character, user query.}

\item{top_k}{A numeric, number of top K records to return.}

\item{verbose}{A logical. If \code{FALSE} you won't be getting
additional diagnostic messages.}
}
\description{
Finding top K records in a vector database.
}
