% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitAI-package.R
\docType{package}
\name{GitAI-package}
\alias{GitAI}
\alias{GitAI-package}
\title{Derive knowledge from GitHub or GitLab repositories with the use of AI/LLM}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Scan multiple 'Git' repositories, pull specified files content and process it with large language models. You can summarize the content in specific way, extract information and data, or find answers to your questions about the repositories. The output can be stored in vector database and used for semantic search or as a part of a RAG (Retrieval Augmented Generation) prompt.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-world-devs/GitAI}
}

}
\author{
\strong{Maintainer}: Kamil Wais \email{kamil.wais@gmail.com}

Authors:
\itemize{
  \item Krystian Igras \email{krystian8207@gmail.com}
  \item Maciej Banas \email{banasmaciek@gmail.com}
}

}
