% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{is_verbose}
\alias{is_verbose}
\title{Checks if \code{GitAI} is verbose}
\usage{
is_verbose()
}
\value{
A logical. Default is \code{TRUE}.
}
\description{
The function searches for \code{GITAI_VERBOSE} environmental variable
and tries to read it value as logical.
}
