% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitAI-package.R
\name{missing_deps_note_fix}
\alias{missing_deps_note_fix}
\title{This function is meant to fix 'Namespaces in Imports field not imported from:' R check note.
The note shows up when namespace is used to create package object (not function) or
within file marked at '.Rbuildignore' file.}
\usage{
missing_deps_note_fix()
}
\description{
This function is meant to fix 'Namespaces in Imports field not imported from:' R check note.
The note shows up when namespace is used to create package object (not function) or
within file marked at '.Rbuildignore' file.
}
