% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_database.R
\name{set_database}
\alias{set_database}
\title{Setting database in \code{GitAI} object.}
\usage{
set_database(gitai, provider = "Pinecone", ...)
}
\arguments{
\item{gitai}{A \code{GitAI} object.}

\item{provider}{A string. Name of database provider.}

\item{...}{Additional arguments to pass to database provider constructor.}
}
\description{
Setting database in \code{GitAI} object.
}
