% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repos.R
\name{set_github_repos}
\alias{set_github_repos}
\title{Set GitHub repositories in \code{GitAI} object.}
\usage{
set_github_repos(gitai, host = NULL, repos, verbose = is_verbose())
}
\arguments{
\item{gitai}{A \code{GitAI} object.}

\item{host}{A character, GitHub host.}

\item{repos}{A character vector or repositories full names.}

\item{verbose}{A logical. If \code{FALSE} you won't be getting
additional diagnostic messages.}
}
\value{
A \code{GitAI} object.
}
\description{
Set GitHub repositories in \code{GitAI} object.
}
