% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_llm.R
\name{set_llm}
\alias{set_llm}
\alias{get_llm_defaults}
\title{Set Large Language Model in \code{GitAI} object.}
\usage{
set_llm(gitai, provider = "openai", ...)

get_llm_defaults(provider)
}
\arguments{
\item{gitai}{A \code{GitAI} object.}

\item{provider}{Name of LLM provider, a string. Results with setting up LLM using
\code{ellmer::chat_<provider>} function.}

\item{...}{Other arguments to pass to corresponding \code{ellmer::chat_<provider>} function.
Please use \link{get_llm_defaults} to get default model arguments.}
}
\value{
A \code{GitAI} object.
}
\description{
Set Large Language Model in \code{GitAI} object.
}
