% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_llm.R
\name{set_prompt}
\alias{set_prompt}
\title{Set prompt.}
\usage{
set_prompt(gitai, system_prompt)
}
\arguments{
\item{gitai}{A \code{GitAI} object.}

\item{system_prompt}{A system prompt.}
}
\value{
A \code{GitAI} object.
}
\description{
Set prompt.
}
