% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_repos.R
\name{process_repos}
\alias{process_repos}
\title{Run LLM on \code{GitAI} repositories content}
\usage{
process_repos(gitai, depth = 1, verbose = is_verbose())
}
\arguments{
\item{gitai}{A \code{GitAI} object.}

\item{depth}{A numeric, maximum depth of folders to process.}

\item{verbose}{A logical. If \code{FALSE} you won't be getting
additional diagnostic messages.}
}
\value{
A list.
}
\description{
Run LLM on \code{GitAI} repositories content
}
