% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_repos.R
\name{process_repos}
\alias{process_repos}
\title{Run LLM on \code{GitAI} repositories content}
\usage{
process_repos(gitai)
}
\arguments{
\item{gitai}{A \code{GitAI} object.}
}
\value{
A list.
}
\description{
Run LLM on \code{GitAI} repositories content
}
