% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repos.R
\name{set_github_repos}
\alias{set_github_repos}
\title{Set GitHub repositories in GitAI object.}
\usage{
set_github_repos(gitai, host = NULL, repos)
}
\arguments{
\item{gitai}{A GitAI object.}

\item{host}{A character, GitHub host.}

\item{repos}{A character vector or repositories full names.}
}
\value{
GitAI object.
}
\description{
Set GitHub repositories in GitAI object.
}
