% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repos.R
\name{set_gitlab_repos}
\alias{set_gitlab_repos}
\title{Set GitLab repositories in GitAI object.}
\usage{
set_gitlab_repos(gitai, host = NULL, repos)
}
\arguments{
\item{gitai}{A GitAI object.}

\item{host}{A character, GitLab host.}

\item{repos}{A character vector or repositories full names.}
}
\value{
GitAI object.
}
\description{
Set GitLab repositories in GitAI object.
}
